projects/metrics-monitoring-alerting-system/requirements.txt
# Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0
django-filter==24.1

# Database and Cache
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0

# Data validation and parsing
pydantic==2.5.3
PyYAML==6.0.1

# Utilities
python-dotenv==1.0.0
celery==5.3.6  # For background tasks (optional)

# Development and testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
faker==22.0.0

# Monitoring and logging
colorlog==6.8.0
